+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## v1.3.4
+
+### Features
+- Added `Changelog` to check latest version response.
+- Added `Critical` parameter.
+
+### Maintenance
+
+- Refactored `List`, `Create`, `Delete` mongodb functions.
+
 ## v1.3.3
 
 ### Important Notes